added validation split

+++ cleartext/models/lstm.py
@@ -12,13 +12,13 @@ def lstm(vocab_size, seq_len, units, weights=None):
                              mask_zero=True)
 
     # encoder
-    enc_in = layers.Input(shape=(seq_len,))
+    enc_in = layers.Input(shape=(seq_len,), dtype='int32')
     enc_embed = embed(enc_in)
     enc_lstm = layers.LSTM(units, return_state=True)
     _, enc_hidden, enc_cell = enc_lstm(enc_embed)
 
     # decoder
-    dec_in = layers.Input(shape=(seq_len,))
+    dec_in = layers.Input(shape=(seq_len,), dtype='int32')
     dec_embed = embed(dec_in)
     dec_lstm = layers.LSTM(units, return_sequences=True, return_state=True)
     dec_out, _, _ = dec_lstm(dec_embed, initial_state=[enc_hidden, enc_cell])
